Add cargo --no-verify test when checking that src dir was not modified while publishing
